requirements.txt

Django
django-bleach==0.6.1
django-cors-headers
django-group-by
django-prometheus==2.1.0
django-python3-ldap
djangorestframework
djangorestframework-simplejwt
psycopg2
pyasn1
PyJWT
python-dateutil
pytz
requests
sqlparse
urllib3
virtualenv
virtualenvwrapper-win

